requirements.txt
diagrams
PyYAML